+++ examples/site/get_sites.py
@@ -1,0 +1,13 @@
+import os
+
+import ornitho
+
+ornitho.consumer_key = os.environ.get("ORNITHO_CONSUMER_KEY")
+ornitho.consumer_secret = os.environ.get("ORNITHO_CONSUMER_SECRET")
+ornitho.user_email = os.environ.get("ORNITHO_USER_EMAIL")
+ornitho.user_pw = os.environ.get("ORNITHO_USER_PW")
+ornitho.api_base = os.environ.get("ORNITHO_API_BASE")
+
+sites = ornitho.Site.get_sites([18287, 12344])
+for site in sites:
+    print(f"Site with ID {site.id_}: {site.custom_name} / {site.local_name}")